+++ frontend/src/App.jsx
@@ -21,7 +21,7 @@ const ProtectedRoutes = () => {
 };
 
 function App() {
-  const API_URL = 'http://127.0.0.1:5001';
+  const API_URL = 'https://myapp-backend-y8f2.onrender.com';
   const [userId, setUserId] = useState(() => localStorage.getItem('userId'));
   useEffect(() => {
     if (userId) localStorage.setItem('userId', userId);
